fix(bundle): wrap concatenated files in a bundle
Previously, since there was a top-level declaration of `var Reflect`,
this was causing a condition of `if 'Reflect' in window` to be true,
which was causing an error in Safari when Object.defineProperty
was being called on window for the `Reflect` property. This fix
wraps the concatenated files in a self-executing function so that 
no global variables are accidentally declared.